// src/app/models/interfaces/dashboard.interface.ts
export interface IExpensesChartClosingBalanceResponse {
  lable?: IChartLabelType;
  chartTitle?: string;
  operatingcostActiveyear?: ClosingBalanceResponse;
  operatingcostLastyear?: ClosingBalanceResponse;
  indirectexpensesActiveyear?: ClosingBalanceResponse;
  indirectexpensesLastyear?: ClosingBalanceResponse;
}

export interface IRevenueChartClosingBalanceResponse {
  lable?: IChartLabelType;
  chartTitle?: string;
  revenuefromoperationsActiveyear?: ClosingBalanceResponse;
  revenuefromoperationsLastyear?: ClosingBalanceResponse;
  otherincomeActiveyear?: ClosingBalanceResponse;
  otherincomeLastyear?: ClosingBalanceResponse;
}
// ...
export enum ChartFilterType {
  ThisMonthToDate,
  ThisQuarterToDate,
  ThisFinancialYearToDate,
  ThisYearToDate,
  LastMonth,
  LastQuater,
  LastFiancialYear,
  LastYear,
  Custom
}

export enum ChartType {
  Revenue,
  Expense,
  ProfitLoss
}

export interface IChartLabelType {
  activeYearLabel?: string;
  lastYearLabel?: string;
}